requirements.txt
av
mediapipe
numpy 
opencv-contrib-python
opencv-python
Pillow 
playsound 
streamlit 
streamlit-webrtc
